Car asociate with client

+++ src/main/java/com/rentsclients/rentsandclients/service/CarService.java
@@ -6,7 +6,6 @@ import com.rentsclients.rentsandclients.Entity.ClientEntity;
 import com.rentsclients.rentsandclients.Exceptions.ClientNotFoundException;
 import com.rentsclients.rentsandclients.Repository.CarRepository;
 import com.rentsclients.rentsandclients.Repository.ClientRepository;
-import com.rentsclients.rentsandclients.mapper.CarMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,78 +14,98 @@ import java.util.List;
 public class CarService {
     private final CarRepository carRepository;
 
-    private final CarMapper carMap;
-
     private final ClientRepository clientRepository;
 
-    public CarService(CarRepository carRepository, CarMapper carMap, ClientRepository clientRepository) {
+    public CarService(CarRepository carRepository, ClientRepository clientRepository) {
         this.carRepository = carRepository;
-        this.carMap = carMap;
         this.clientRepository = clientRepository;
     }
 
-    public void createACar(CarDTO carDTO) {
+    public CarDTO createACar(CarDTO carDTO) {
         if (carDTO == null) {
             throw new IllegalArgumentException("CarDTO cannot be null");
         }
 
-        CarEntity carEntity = carMap.toCarEntity(carDTO);
+        CarEntity converterInEntity = new CarEntity();
+        converterInEntity.setCarid(carDTO.getCarid());
+        converterInEntity.setBrand(carDTO.getBrand());
+        converterInEntity.setModel(carDTO.getModel());
+        converterInEntity.setPlate(carDTO.getPlate());
+        converterInEntity.setActivated(carDTO.isActivated());
 
-        if (carRepository.existsByPlate(carEntity.getPlate())){
+        if (carRepository.existsByPlate(converterInEntity.getPlate())) {
             throw new RuntimeException("The plate is already registered");
         }
 
-        carRepository.save(carEntity);
-    }
+        CarEntity carSaved = carRepository.save(converterInEntity);
 
-    //por alugm motivo o dto nao está pegando o id por isso estou passando aqui o car entity
-    public List<CarEntity> getAllCars() {
-        List<CarEntity> allCars = carRepository.findAll();
-//        List<CarDTO> cars = carMapper.ToCarsDtoList(allCars);
-        return allCars;
-    }
-
-    public CarDTO getASpecifiqueCarID(Long id) {
-        return carMap.toCarDto(carRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("CarID not found")));
+        return new CarDTO(
+                carSaved.getCarid(),
+                carSaved.getBrand(),
+                carSaved.getModel(),
+                carSaved.getPlate(),
+                carSaved.isActivated(),
+                null
+        );
     }
 
     public CarDTO updateCarByID(Long id, CarDTO carDTO) {
         CarEntity findCar = carRepository.findById(id).orElseThrow(() -> new RuntimeException("CarID not found"));
 
-        CarEntity updateCar = carMap.toCarEntity(carDTO);
+        findCar.setBrand(carDTO.getBrand());
+        findCar.setModel(carDTO.getModel());
+        findCar.setPlate(carDTO.getPlate());
+        findCar.setActivated(carDTO.isActivated());
 
-        if (carRepository.existsByPlate(updateCar.getPlate())){
+        if (carRepository.existsByPlate(findCar.getPlate())) {
             throw new RuntimeException("The plate is already registered ");
         }
-        updateCar.setCarid(findCar.getCarid());
 
-        CarEntity savedCar = carRepository.save(updateCar);
-        return carMap.toCarDto(savedCar);
+        CarEntity savedCar = carRepository.save(findCar);
+
+        return new CarDTO(
+                savedCar.getCarid(),
+                savedCar.getBrand(),
+                savedCar.getModel(),
+                savedCar.getPlate(),
+                savedCar.isActivated(),
+                null
+        );
     }
 
-    public CarDTO activateOrDeactivateCarByID(long id, CarDTO carDTO) {
+    public void activateOrDeactivateCarByID(long id, CarDTO carDTO) {
         CarEntity findCar = carRepository.findById(id).orElseThrow(() -> new RuntimeException("ID not found. The car will not be updated"));
 
         findCar.setActivated(carDTO.isActivated());
-        CarEntity savedCar = carRepository.save(findCar);
-
-        return carMap.toCarDto(savedCar);
+        carRepository.save(findCar);
     }
 
-    public CarDTO associateCarWithClient(long carId, Long userId) {
+
+    public void associateCarWithClient(long carId, Long userId) {
         CarEntity findCar = carRepository.findById(carId).orElseThrow(() -> new RuntimeException("ID not found. The car will not be updated"));
 
         ClientEntity findUser = clientRepository.findById(userId).orElseThrow(() -> new ClientNotFoundException("Client not found. The car will not be updated"));
 
         findCar.setClient(findUser);
 
-        CarEntity savedCar = carRepository.save(findCar);
-
-        return carMap.toCarDto(savedCar);
+        carRepository.save(findCar);
     }
 
+    public List<CarEntity> getAllCars() {
+        return carRepository.findAll();
+    }
+
+//
+//    public CarDTO getASpecifiqueCarID(Long id) {
+//        return carMap.toCarDto(carRepository.findById(id)
+//                .orElseThrow(() -> new RuntimeException("CarID not found")));
+//    }
+
     public void deleteCarByID(Long id) {
+        List<CarEntity> findCars = carRepository.findAll();
+        if (findCars.isEmpty()) {
+            throw new RuntimeException("Car not found cannot be deleted");
+        }
         carRepository.deleteById(id);
     }
 }
